server/requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.6
dnspython==2.4.2
Flask==2.3.2
flask[async]==2.3.2
Flask-RESTful==0.3.10
Flask-Cors==3.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mongoengine==0.27.0
PyJWT==2.8.0
pymongo==4.4.1
Werkzeug==2.3.6
httpx~=0.24.1